Add starry background and interactive globe tooltips
