+++ app/globals.css
@@ -69,21 +69,51 @@ html {
 
 body {
   font-family: var(--font-primary);
-  background: var(--warm-white);
+  background: radial-gradient(circle at 50% 20%, rgba(241, 196, 15, 0.15) 0%, transparent 60%),
+    linear-gradient(135deg, #0d1b2a 0%, #1b263b 50%, #415a77 100%);
+  background-size: 100% 100%, 400% 400%;
+  animation: gradientShift 25s ease infinite;
   color: var(--charcoal);
   line-height: 1.6;
   overflow-x: hidden;
 }
 
+body::before {
+  content: '';
+  position: fixed;
+  inset: 0;
+  pointer-events: none;
+  background: radial-gradient(white 1px, transparent 1px);
+  background-size: 3px 3px;
+  opacity: 0.15;
+  animation: starScroll 60s linear infinite;
+  z-index: -1;
+}
+
+@keyframes gradientShift {
+  0% {
+    background-position: 0% 50%;
+  }
+  50% {
+    background-position: 100% 50%;
+  }
+  100% {
+    background-position: 0% 50%;
+  }
+}
+
+@keyframes starScroll {
+  from {
+    transform: translateY(0);
+  }
+  to {
+    transform: translateY(-1000px);
+  }
+}
+
 /* App Container */
 .app-container {
   min-height: 100vh;
-  background: linear-gradient(
-    180deg,
-    var(--warm-white) 0%,
-    var(--soft-gray) 50%,
-    var(--cream) 100%
-  );
   position: relative;
 }
 
@@ -614,6 +644,32 @@ body {
   }
 }
 
+/* Tooltip Styles */
+.tooltip {
+  position: fixed;
+  pointer-events: none;
+  background: rgba(0, 0, 0, 0.7);
+  color: #fff;
+  padding: 4px 8px;
+  border-radius: 4px;
+  font-size: 0.8rem;
+  white-space: nowrap;
+  z-index: 200;
+  transform: translate(-50%, -120%);
+  animation: tooltipFade var(--transition-medium);
+}
+
+@keyframes tooltipFade {
+  from {
+    opacity: 0;
+    transform: translate(-50%, -100%) scale(0.95);
+  }
+  to {
+    opacity: 1;
+    transform: translate(-50%, -120%) scale(1);
+  }
+}
+
 /* High contrast mode support */
 @media (prefers-contrast: high) {
   :root {
@@ -653,16 +709,7 @@ body {
     --cream: #1e1e1e;
     --charcoal: #e0e0e0;
   }
-  
-  .app-container {
-    background: linear-gradient(
-      180deg,
-      #1a1a1a 0%,
-      #2a2a2a 50%,
-      #1e1e1e 100%
-    );
-  }
-  
+
   .comparison-item {
     background: rgba(42, 42, 42, 0.8);
     border-color: rgba(167, 201, 87, 0.3);
