+++ app/page.tsx
@@ -1,120 +1,248 @@
 'use client'
 
-import { useState, useEffect } from 'react'
+import React, { Suspense, useEffect, useRef, useState } from 'react'
+import * as THREE from 'three'
+import type { GlobeMethods } from 'react-globe.gl'
 import ImageComparison from './components/ImageComparison'
 
-interface ImagePair {
-  id: string
-  title: string
-  description: string
-  beforeImage: {
-    src: string
-    alt: string
-    label: string
-  }
-  afterImage: {
-    src: string
-    alt: string
-    label: string
+const Globe = React.lazy(() => import('react-globe.gl'))
+
+interface LocationPoint {
+  lat: number
+  lng: number
+  size?: number
+  color?: string
+  name: string
+  images: {
+    before: { src: string; alt: string; label: string }
+    after: { src: string; alt: string; label: string }
   }
 }
 
-const sampleImagePairs: ImagePair[] = [
+const locations: LocationPoint[] = [
   {
-    id: '1',
-    title: 'Urban Transformation',
-    description: 'From concrete jungle to green sanctuary - reimagining city spaces with vertical gardens and sustainable architecture.',
-    beforeImage: {
-      src: '/images/urban/bellagio.png',
-      alt: 'Urban cityscape with concrete buildings and pollution',
-      label: 'Present Day'
+    lat: 52.2297,
+    lng: 21.0122,
+    size: 0.35,
+    color: '#ff8c00',
+    name: 'Warsaw',
+    images: {
+      before: { src: '/images/urban/pawia.png', alt: 'Warsaw today', label: 'Today' },
+      after: { src: '/images/urban/pawia-punk.webp', alt: 'Solarpunk Warsaw', label: 'Solarpunk' },
     },
-    afterImage: {
-      src: '/images/urban/bellagio2.png',
-      alt: 'Green urban landscape with vertical gardens and sustainable buildings',
-      label: 'Solarpunk Future'
-    }
   },
   {
-    id: '2',
-    title: 'Energy Revolution',
-    description: 'Transitioning from fossil fuel dependency to clean, renewable energy sources integrated harmoniously with nature.',
-    beforeImage: {
-      src: '/images/urban/vegas1.jpg',
-      alt: 'Industrial landscape with smokestacks and pollution',
-      label: 'Industrial Age'
+    lat: 36.1126,
+    lng: -115.1767,
+    name: 'Las Vegas Strip',
+    images: {
+      before: { src: '/images/urban/bellagio.png', alt: 'Las Vegas Strip today', label: 'Today' },
+      after: { src: '/images/urban/bellagio2.png', alt: 'Solarpunk Las Vegas Strip', label: 'Solarpunk' },
     },
-    afterImage: {
-      src: '/images/urban/vegas2.png',
-      alt: 'Solar panels and wind turbines in a green landscape',
-      label: 'Renewable Future'
-    }
   },
   {
-    id: '3',
-    title: 'Transportation Evolution',
-    description: 'Moving from car-dependent infrastructure to sustainable, community-focused transportation systems.',
-    beforeImage: {
-      src: '/images/urban/eifel.webp',
-      alt: 'Highway filled with cars and traffic',
-      label: 'Car-Centric'
+    lat: 36.1699,
+    lng: -115.1398,
+    name: 'Downtown Las Vegas',
+    images: {
+      before: { src: '/images/urban/vegas1.jpg', alt: 'Downtown Las Vegas today', label: 'Today' },
+      after: { src: '/images/urban/vegas2.png', alt: 'Solarpunk Downtown Las Vegas', label: 'Solarpunk' },
     },
-    afterImage: {
-      src: '/images/urban/eifel2.png',
-      alt: 'Bike lanes and green transportation corridors',
-      label: 'Sustainable Mobility'
-    }
   },
-
+  {
+    lat: 48.8566,
+    lng: 2.3522,
+    name: 'Paris',
+    images: {
+      before: { src: '/images/urban/eifel.webp', alt: 'Paris today', label: 'Today' },
+      after: { src: '/images/urban/eifel2.png', alt: 'Solarpunk Paris', label: 'Solarpunk' },
+    },
+  },
 ]
 
-export default function Home() {
-  const [isLoaded, setIsLoaded] = useState(false)
+export default function GlobePage() {
+  const globeRef = useRef<GlobeMethods | undefined>(undefined)
+  const [isMounted, setIsMounted] = useState(false)
+  const [viewport, setViewport] = useState<{ width: number; height: number }>({ width: 0, height: 0 })
+  const [selected, setSelected] = useState<LocationPoint | null>(null)
+  const [hovered, setHovered] = useState<LocationPoint | null>(null)
+  const [mousePos, setMousePos] = useState<{ x: number; y: number }>({ x: 0, y: 0 })
 
   useEffect(() => {
-    setIsLoaded(true)
+    setIsMounted(true)
+    const computeViewport = () => {
+      const width = typeof window !== 'undefined' ? window.innerWidth : 800
+      const height = typeof window !== 'undefined' ? Math.round(window.innerHeight * 0.7) : 600
+      setViewport({ width, height })
+    }
+    computeViewport()
+    window.addEventListener('resize', computeViewport)
+    return () => window.removeEventListener('resize', computeViewport)
   }, [])
 
+  useEffect(() => {
+    if (!isMounted) return
+    const canvas = globeRef.current?.renderer().domElement
+    if (!canvas) return
+    const handleMove = (e: MouseEvent) => {
+      setMousePos({ x: e.clientX, y: e.clientY })
+    }
+    canvas.addEventListener('mousemove', handleMove)
+    return () => canvas.removeEventListener('mousemove', handleMove)
+  }, [isMounted])
+
+  useEffect(() => {
+    if (!isMounted) return
+    let canceled = false
+    const tryInit = () => {
+      if (canceled) return
+      const globe = globeRef.current
+      if (!globe) {
+        requestAnimationFrame(tryInit)
+        return
+      }
+      globe.pointOfView({ lat: 30, lng: 10, altitude: 2.2 }, 1500)
+      const controls = globe.controls()
+      if (controls) {
+        controls.autoRotate = true
+        controls.autoRotateSpeed = 0.35
+      }
+    }
+    tryInit()
+    return () => {
+      canceled = true
+    }
+  }, [isMounted])
+
+  // Add semi-transparent clouds layer
+  useEffect(() => {
+    if (!isMounted) return
+    let canceled = false
+    let animationFrameId = 0
+    let cloudsMesh: THREE.Mesh | null = null
+    const CLOUDS_IMG_URL = '/fair_clouds_4k.png'
+    const CLOUDS_ALT = 0.004
+    const CLOUDS_ROTATION_SPEED = -0.006 // deg/frame
+
+    const initClouds = () => {
+      if (canceled) return
+      const globe = globeRef.current
+      if (!globe) {
+        requestAnimationFrame(initClouds)
+        return
+      }
+      const loader = new THREE.TextureLoader()
+      loader.load(CLOUDS_IMG_URL, (cloudsTexture) => {
+        cloudsTexture.colorSpace = THREE.SRGBColorSpace
+        cloudsMesh = new THREE.Mesh(
+          new THREE.SphereGeometry(globe.getGlobeRadius() * (1 + CLOUDS_ALT), 75, 75),
+          new THREE.MeshPhongMaterial({ map: cloudsTexture, transparent: true, depthWrite: false })
+        )
+        globe.scene().add(cloudsMesh)
+
+        const rotateClouds = () => {
+          if (canceled) return
+          if (cloudsMesh) {
+            cloudsMesh.rotation.y += (CLOUDS_ROTATION_SPEED * Math.PI) / 180
+          }
+          animationFrameId = requestAnimationFrame(rotateClouds)
+        }
+        rotateClouds()
+      })
+    }
+    initClouds()
+
+    return () => {
+      canceled = true
+      if (animationFrameId) cancelAnimationFrame(animationFrameId)
+      const globe = globeRef.current
+      if (cloudsMesh && globe) {
+        try {
+          globe.scene().remove(cloudsMesh)
+        } catch {}
+      }
+    }
+  }, [isMounted])
+
   return (
-    <div className={`page-container ${isLoaded ? 'loaded' : ''}`}>
-      <div className="intro-section">
-        <div className="intro-content">
-          <h2 className="intro-title">Envisioning Tomorrow</h2>
-          <p className="intro-text">
-            PLACEHOLDER: Explore the transformation from our current world to a sustainable, 
-            equitable solarpunk future. Drag the slider to reveal possibilities 
-            where technology and nature exist in harmony.
-          </p>
-        </div>
-        <div className="decorative-elements">
-          <div className="floating-leaf leaf-1">🍃</div>
-          <div className="floating-leaf leaf-2">🌱</div>
-          <div className="floating-leaf leaf-3">🌿</div>
-        </div>
-      </div>
-
-      <div className="comparison-feed">
-        {sampleImagePairs.map((pair, index) => (
-          <div key={pair.id} className={`comparison-item item-${index + 1}`}>
-            <div className="comparison-header">
-              <h3 className="comparison-title">{pair.title}</h3>
-              <p className="comparison-description">{pair.description}</p>
-            </div>
-            <ImageComparison
-              beforeImage={pair.beforeImage}
-              afterImage={pair.afterImage}
+    <div style={{ position: 'relative' }}>
+      <div style={{ width: '100%', height: viewport.height }}>
+        {isMounted && (
+          <Suspense fallback={null}>
+            <Globe
+              ref={globeRef}
+              width={viewport.width}
+              height={viewport.height}
+              backgroundColor="rgba(0, 0, 0, 0)"
+              globeImageUrl="https://unpkg.com/three-globe/example/img/earth-blue-marble.jpg"
+              bumpImageUrl="https://unpkg.com/three-globe/example/img/earth-topology.png"
+              pointsData={locations}
+              pointAltitude={(p: any) =>
+                ((p as LocationPoint).name === 'Warsaw' ? 0.12 : 0.05) +
+                0.01 * Math.sin(Date.now() / 200)
+              }
+              pointColor={(p: any) => (p as LocationPoint).color || '#ffe066'}
+              pointRadius={(p: any) =>
+                ((p as LocationPoint).size || 0.3) + 0.02 * Math.sin(Date.now() / 200)
+              }
+              onPointClick={(p) => setSelected(p as LocationPoint)}
+              onPointHover={(p) => setHovered(p as LocationPoint | null)}
             />
-          </div>
-        ))}
+          </Suspense>
+        )}
       </div>
 
-      <div className="footer-section">
-        <div className="footer-content">
-          <p className="footer-text">
-            🌍 Building sustainable futures through imagination and action
-          </p>
+      {selected && (
+        <div
+          onClick={() => setSelected(null)}
+          style={{
+            position: 'fixed',
+            inset: 0,
+            background: 'rgba(0,0,0,0.45)',
+            display: 'flex',
+            alignItems: 'center',
+            justifyContent: 'center',
+            zIndex: 50,
+          }}
+        >
+          <div
+            onClick={(e) => e.stopPropagation()}
+            style={{
+              background: 'white',
+              color: '#111',
+              borderRadius: 12,
+              boxShadow: '0 10px 30px rgba(0,0,0,0.2)',
+              width: 'min(90vw, 560px)',
+              padding: 24,
+            }}
+          >
+            <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: 12 }}>
+              <h2 style={{ fontSize: 20, margin: 0 }}>{selected.name}</h2>
+              <button
+                onClick={() => setSelected(null)}
+                style={{
+                  border: 'none',
+                  background: 'transparent',
+                  fontSize: 22,
+                  lineHeight: 1,
+                  cursor: 'pointer',
+                }}
+                aria-label="Close"
+              >
+                ×
+              </button>
+            </div>
+            <ImageComparison beforeImage={selected.images.before} afterImage={selected.images.after} />
+          </div>
         </div>
-      </div>
+      )}
+
+      {hovered && (
+        <div className="tooltip" style={{ left: mousePos.x, top: mousePos.y }}>
+          {hovered.name}
+        </div>
+      )}
     </div>
   )
 }
+
